+++ docs/articles/get_started_files/sigmajs-binding-0.0.3/sigmajs.js
@@ -824,6 +824,84 @@ if (HTMLWidgets.shinyMode) {
 			}
 		}
 	);
+	
+	// filter greater than
+	Shiny.addCustomMessageHandler('sg_filter_gt_p',
+		function (message) {
+			var s = get_sigma_graph(message.id);
+			
+			if (typeof s != 'undefined') {
+			  
+			  var filter = new sigma.plugins.filter(s);
+			  
+			  if(message.target === "both"){
+          filter
+            .undo()
+            .nodesBy(function(n) {
+              return n[message.var] > message.input;
+            })
+            .edgesBy(function(e) {
+              return e[message.var] > message.input;
+            })
+            .apply();
+			  } else if(message.target === "nodes"){
+          filter
+            .undo()
+            .nodesBy(function(n) {
+              return n[message.var] > message.input;
+            })
+            .apply();
+			  } else {
+          filter
+            .undo()
+            .edgesBy(function(e) {
+              return e[message.var] > message.input;
+            })
+            .apply();
+			  }
+				
+			}
+		}
+	);
+	
+	// filter less than
+	Shiny.addCustomMessageHandler('sg_filter_lt_p',
+		function (message) {
+			var s = get_sigma_graph(message.id);
+			
+			if (typeof s != 'undefined') {
+			  
+			  var filter = new sigma.plugins.filter(s);
+			  
+			  if(message.target === "both"){
+          filter
+            .undo()
+            .nodesBy(function(n) {
+              return n[message.var] < message.input;
+            })
+            .edgesBy(function(e) {
+              return e[message.var] < message.input;
+            })
+            .apply();
+			  } else if(message.target === "nodes"){
+          filter
+            .undo()
+            .nodesBy(function(n) {
+              return n[message.var] < message.input;
+            })
+            .apply();
+			  } else {
+          filter
+            .undo()
+            .edgesBy(function(e) {
+              return e[message.var] < message.input;
+            })
+            .apply();
+			  }
+				
+			}
+		}
+	);
 
 }